Refactoring Fhir to HV Date conversion
Adding extension method for converting fhir element to HealthServiceDateTime
Using newly added extension for calculating "when" for various observations
Adding exceptions where date is not present but required by Health Vault
Renaming FhirToApproximateDateTime.cs to a more generic name FhirtoHealthVaultDateTime
Moving generic methods out of ObservationToHealthVault.cs
