+++ Microsoft.HealthVault.Fhir/Transformers/ObservationToBloodGlucose.cs
@@ -20,7 +20,11 @@ namespace Microsoft.HealthVault.Fhir.Transformers
             var bloodGlucose = observation.ToThingBase<BloodGlucose>();
 
             bloodGlucose.Value = ObservationToHealthVault.GetThingValueFromQuantity<BloodGlucoseMeasurement>(observation.Value as Quantity);
-            bloodGlucose.When = ObservationToHealthVault.GetHealthVaultTimeFromEffectiveDate(observation.Effective);
+
+            if (observation.Effective == null)
+                throw new ArgumentException("Effective is required");
+
+            bloodGlucose.When = observation.Effective.ToHealthServiceDateTime();
 
             var bloodGlucoseExtension = observation.GetExtension(HealthVaultExtensions.BloodGlucose);
