+++ Microsoft.HealthVault.Fhir/Transformers/ObservationToHeight.cs
@@ -6,6 +6,7 @@
 //
 // THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
 using Hl7.Fhir.Model;
 using Microsoft.HealthVault.ItemTypes;
 
@@ -18,7 +19,10 @@ namespace Microsoft.HealthVault.Fhir.Transformers
             var height = observation.ToThingBase<ItemTypes.Height>();
 
             height.Value = ObservationToHealthVault.GetThingValueFromQuantity<Length>(observation.Value as Quantity);
-            height.When = ObservationToHealthVault.GetHealthVaultTimeFromEffectiveDate(observation.Effective);
+
+            if (observation.Effective == null)
+                throw new ArgumentException("Effective is required");
+            height.When = observation.Effective.ToHealthServiceDateTime();
 
             return height;
         }
