+++ Microsoft.HealthVault.Fhir/Transformers/ObservationToWeight.cs
@@ -6,6 +6,7 @@
 //
 // THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System;
 using Hl7.Fhir.Model;
 using Microsoft.HealthVault.ItemTypes;
 
@@ -18,7 +19,10 @@ namespace Microsoft.HealthVault.Fhir.Transformers
             var weight = observation.ToThingBase<ItemTypes.Weight>();
 
             weight.Value = ObservationToHealthVault.GetThingValueFromQuantity<WeightValue>(observation.Value as Quantity);
-            weight.When = ObservationToHealthVault.GetHealthVaultTimeFromEffectiveDate(observation.Effective);
+            if (observation.Effective == null)
+                throw new ArgumentException("Effective is required");
+
+            weight.When = observation.Effective.ToHealthServiceDateTime();
 
             return weight;
         }
